rename `prisma` client instance
